Forward file data to vfile and back to vinyl
gulp-data standardized saving custom data to the `data` property, so
we're updating convert-vinyl-to-vfile because the latest version
forwards vinyl data to vfile for any unified plugin which might use it,
like rehype-meta, and after processing we're saving it back to the vinyl
file in case any unified plugin modified it.

One use case is for parsing and saving data from remark-frontmatter to
make it available to gulp plugins down the stream.